Solution to Input and Print a Double

+++ Input and Print a Double@@6788d5012db25efec946f942/code.c
@@ -1,7 +1,7 @@
 #include <stdio.h>
 int main()
 {
-    int float num;
+    float num;
     scanf("%f",&num);
     printF("You entered: %f",num);
     return 0;
